% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(exampleData)
}
\arguments{
\item{exampleData}{String. These example data are included: mRNA_path, microRNA_path, drug.ic50, drug.r, drug.fdr, drug.info, drug.edgelist, breast_cancer and brc_candidates.}
}
\description{
This function is used to get example data.
}
