% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorCD-package.R
\docType{package}
\name{PriorCD-package}
\alias{PriorCD-package}
\alias{PriorCD}
\title{Prioritizing cancer drugs for interested cancer}
\description{
This package implements methods to predict priorities of therapeutic drugs against interested cancer by combining drug functional similarity network and global network propagation algorithm. Besides, users can validate the prioritizing results and visualize the network structure of the resultant drugs.
}
